Updates exchange balances layout
[skip ci]
